Update filter bindings to display on screen
Update the controller (filter_controller.go) to set up the binding to be
didsplayed. Additionally, update the translations for the binding name
to be shorter, to make it more likely the text will fit at the bottom of
the screen. This also involved updating the keybindings_*.md files with
go generate ./...

The keybinding screen in particular (which is a filterable popup) seems
frustrating for users. In #3141, @kadaan suggested that perhaps the
keybindings page should display a message in the footer about the
presence of this option as a way to increase discoverability.

This change enables this binding displaying on _all_ filterable views.
I'm not sure if it's valuable to restrict this to just the keybinding
view.