Make losing focus on Confirmations count as closing them
This will fix https://github.com/jesseduffield/lazygit/issues/4052
where new users click out of the panel, but the state never gets
updated.

It makes sense to do this as a close and not a confirm, because losing
focus could be due to many things, and doesn't indicate an explicit
confirmation.

It is up to the implementation of the initial popup message that a
cancelling or closing should mean they don't see it again.