Fix discarding submodule changes in nested folders
The current rules for discarding submodule changes is that no other changed item must be also selected. There are some bugs with the current implementation when submodules are in folders.

For example, selecting and discarding a folder with only a nested submodule change will currently do nothing. The submodule changes should be discarded. The folder only contains submodule changes so it should be no different than pressing discard on the submodule entry itself.

Also, I noticed range selecting both the folder and the submodule and then pressing discard would be incorrectly disallowed.